fix highlight

+++ routes/web.php
@@ -13,14 +13,14 @@ Route::get('/login', [AuthController::class, 'login'])->name('login')->middlewar
 Route::post('/login', [AuthController::class, 'loginProcess'])->middleware('guest');
 Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth:web,admin');
 
-Route::get('/news', [NewsController::class, 'index']);
+Route::get('/news', [NewsController::class, 'index'])->middleware('auth:web');
 Route::get('/news/{id}', [NewsController::class, 'show'])->middleware('auth:web,admin');
 
-Route::get('/addNews', [NewsController::class, 'create'])->middleware('auth:admin');
-Route::get('/listNews', [NewsController::class, 'list']);
-Route::get('/news/edit/{id}', [NewsController::class, 'edit']);
-Route::put('/news/edit/{id}', [NewsController::class, 'update']);
-Route::delete('/news/delete/{id}', [NewsController::class, 'destroy']);
+Route::get('/add-news', [NewsController::class, 'create'])->middleware('auth:admin');
+Route::get('/list-news', [NewsController::class, 'list'])->middleware('auth:admin');
+Route::get('/news/edit/{id}', [NewsController::class, 'edit'])->middleware('auth:admin');
+Route::put('/news/edit/{id}', [NewsController::class, 'update'])->middleware('auth:admin');
+Route::delete('/news/delete/{id}', [NewsController::class, 'destroy'])->middleware('auth:admin');
 
 Route::post('/news', [NewsController::class, 'store'])->middleware('auth:admin');
 
@@ -29,6 +29,11 @@ Route::post('/news-comment/{id}/reply', [NewsCommentController::class, 'storeRep
 
 Route::get('/highlight', [VideoController::class, 'index'])->middleware('auth:web');
 Route::get('/highlight/{id}', [VideoController::class, 'show'])->middleware('auth:web');
-Route::get('/addVideo', [VideoController::class, 'create'])->middleware('auth:admin');
+
+Route::get('/add-highlight', [VideoController::class, 'create'])->middleware('auth:admin');
+Route::get('/list-highlight', [VideoController::class, 'list'])->middleware('auth:admin');
+Route::delete('/highlight/delete/{id}', [VideoController::class, 'destroy'])->middleware('auth:admin');
+Route::get('/highlight/edit/{id}', [VideoController::class, 'edit'])->middleware('auth:admin');
+Route::put('/highlight/edit/{id}', [VideoController::class, 'update'])->middleware('auth:admin');
 Route::post('/highlight', [VideoController::class, 'store'])->middleware('auth:admin');
